server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
uvicorn==0.42.0
typing-extensions==4.15.0

# Dependencies identified from imports:
bcrypt
PyJWT
pydantic-settings
motor
google-genai
python-dotenv
pydantic[email]
python-multipart
